Tests
* Corrected namespaces and added throws
* Added Composite header test class
